+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.1.7] (2026-04-16)
 ### Added
 * `pub trait FromJsonStr`
 * `format_version()`: helper to DRY
@@ -18,7 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * typo: `device_pixcel_ratio` to `device_pixel_ratio`
-
 
 ## [0.1.6] (2026-02-12)
 ### Fixed
@@ -54,7 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * first commit
 
-[Unreleased]: https://github.com/aki-akaguma/browserinfo/compare/v0.1.6..HEAD
+[Unreleased]: https://github.com/aki-akaguma/browserinfo/compare/v0.1.7..HEAD
+[0.1.7]: https://github.com/aki-akaguma/browserinfo/compare/v0.1.6..v0.1.7
 [0.1.6]: https://github.com/aki-akaguma/browserinfo/compare/v0.1.5..v0.1.6
 [0.1.5]: https://github.com/aki-akaguma/browserinfo/compare/v0.1.4..v0.1.5
 [0.1.4]: https://github.com/aki-akaguma/browserinfo/compare/v0.1.3..v0.1.4